TSK-1682: Slightly reorganized recruit files for future changes
